WIP: Move test to right testmodule
